Move common request model to a toplevel module
